+++ src/pages/identity-inference/IdentityInferenceResult.tsx
@@ -3,14 +3,17 @@
  * @Author: didadida262
  * @Date: 2024-08-29 10:18:39
  * @LastEditors: didadida262
- * @LastEditTime: 2024-09-02 16:09:57
+ * @LastEditTime: 2024-09-04 18:01:30
  */
 /* eslint-disable @typescript-eslint/restrict-template-expressions */
 /* eslint-disable prettier/prettier */
-
 import { GraphinData } from "@antv/graphin";
+import { notification } from "antd";
 import cn from "classnames";
 import { useEffect, useState } from "react";
+import { AiOutlineLoading3Quarters } from "react-icons/ai";
+import { useParams } from "react-router-dom";
+import { useAsyncFn } from "react-use";
 
 import IdentityInferenceDialogTitle from "@/assets/IdentityInferenceDialogTitle.png";
 import { ButtonCommonV2, EButtonType } from "@/components/ButtonCommonV2";
@@ -24,6 +27,10 @@ import { SpinCommon } from "@/components/SpinCommon";
 import { TableCommonV2 } from "@/components/TableCommonV2";
 import { columns } from "@/services/columns";
 import {
+  detectIdentityRequestType,
+  detectIdentityService
+} from "@/services/detection";
+import {
   generateAddressData,
   generateEdgeTxData,
   initGraphData,
@@ -33,17 +40,20 @@ import pattern from "@/styles/pattern";
 import { IGraphFormData } from "@/utils/IdentityTypes";
 
 export function IdentityInferenceResult() {
+  const { address } = useParams();
   const [pageState, setPageState] = useState("search");
   const [value, setValue] = useState<any>(null);
   const [dataList, setDateList] = useState<any>([]);
-  const [selectedHexData, setSelectedHexData] = useState(initQueryAddress);
+  const [selectedHexData, setSelectedHexData] = useState(
+    address || initQueryAddress
+  );
   const [graphData, setGraphData] = useState<GraphinData>(initGraphData);
-  const [loading, setloading] = useState(false);
+  const [loading, setloading] = useState(true);
 
   const [formData, setFormData] = useState<IGraphFormData>({
     date: ["0", "latest"],
     tokenType: "ETH",
-    address: initQueryAddress
+    address: address || initQueryAddress
   });
 
   const handleClickGraphin = (
@@ -54,9 +64,20 @@ export function IdentityInferenceResult() {
       setSelectedHexData(hexString);
     }
   };
+  const start = async () => {
+    setloading(true);
+    try {
+      const params: detectIdentityRequestType = {
+        address: address || "",
+        chain: "eth"
+      };
+      const respose = await detectIdentityService(params);
+      console.log("respose>>>", respose);
 
-  useEffect(() => {
-    // 请求
+      setloading(false);
+    } catch (error) {
+      setloading(false);
+    }
     const res = [
       {
         name: "测试数据",
@@ -85,10 +106,25 @@ export function IdentityInferenceResult() {
       }
     ];
     setDateList(res);
+  };
+
+  useEffect(() => {
+    // setSelectedHexData(address || initQueryAddress);
+    void start();
   }, []);
 
   return loading
-    ? <SpinCommon className=" w-full h-full" />
+    ? <div
+        className={cn(
+          "w-full h-full absolute top-0 left-0",
+          `${pattern.flexCenter}`
+        )}
+      >
+        <AiOutlineLoading3Quarters
+          className="ml-2 animate-spin"
+          style={{ color: "white", fontSize: "24px" }}
+        />
+      </div>
     : <div
         className={cn(
           " w-full h-full pt-[0px] fadeIn gap-y-6 flex flex-col  relative"
